+++ src/common/Utilities/Containers.h
@@ -186,31 +186,6 @@ namespace Trinity
             return itr != map.end() ? AddressOrSelf(itr->second) : nullptr;
         }
 
-        /**
-         * @class IteratorPair
-         *
-         * @brief Utility class to enable range for loop syntax for multimap.equal_range uses
-         */
-        template<class iterator>
-        class IteratorPair
-        {
-        public:
-            IteratorPair() : _iterators() { }
-            IteratorPair(std::pair<iterator, iterator> iterators) : _iterators(iterators) { }
-
-            iterator begin() const { return _iterators.first; }
-            iterator end() const { return _iterators.second; }
-
-        private:
-            std::pair<iterator, iterator> _iterators;
-        };
-
-        template<class M>
-        inline auto MapEqualRange(M& map, typename M::key_type const& key) -> IteratorPair<decltype(map.begin())>
-        {
-            return { map.equal_range(key) };
-        }
-
         template<class K, class V, template<class, class, class...> class M, class... Rest>
         void MultimapErasePair(M<K, V, Rest...>& multimap, K const& key, V const& value)
         {
